Refactor theme
