+++ src/ui/theme.js
@@ -1,0 +1,14 @@
+const theme = {
+  global: {
+    colors: {
+    brand: '#f44283'
+    },
+    font: {
+      family: 'Roboto',
+      size: '14px',
+      height: '20px'
+    },
+  },
+};
+
+export default theme;
